add bazar

+++ src/Api_FreeSound/apifreesound.py
@@ -51,31 +51,35 @@ class apifreesound():
         return reponse.json()['results']
 
     def dl_son(self, id, HQ=False):
-
-        payload = {'token': self.cleAPI}
-        reponse = requests.get(
-            f'{self.url}/apiv2/sounds/{id}/',
-            params=payload,
-            timeout=1
-            )
-
-        if HQ:
-            url_dl = reponse.json()['previews']['preview-hq-mp3']
-        else:
-            url_dl = reponse.json()['previews']['preview-lq-mp3']
-
-        # Téléchargement du fichier
-        dl_path = gestion_dl.dossier
         
-        reponse = requests.get(url_dl, stream=True)
+        fichier = Path(f"../data/son/{id}.mp3")
 
-        # Écriture du fichier dans le répertoire de destination
-        with open(f'{dl_path}/{id}.mp3', 'wb') as f:
-            for chunk in reponse.iter_content(chunk_size=8192):
-                f.write(chunk)
+        if not fichier.exists():       
+            payload = {'token': self.cleAPI}
+            reponse = requests.get(
+                f'{self.url}/apiv2/sounds/{id}/',
+                params=payload,
+                timeout=1
+                )
 
-        print(f"Fichier téléchargé avec succès dans {dl_path}")
+            if HQ:
+                url_dl = reponse.json()['previews']['preview-hq-mp3']
+            else:
+                url_dl = reponse.json()['previews']['preview-lq-mp3']
 
+            # Téléchargement du fichier
+            dl_path = gestion_dl.dossier
+            
+            reponse = requests.get(url_dl, stream=True)
+
+            # Écriture du fichier dans le répertoire de destination
+            with open(f'{dl_path}/{id}.mp3', 'wb') as f:
+                for chunk in reponse.iter_content(chunk_size=8192):
+                    f.write(chunk)
+
+            print(f"Fichier téléchargé avec succès dans {dl_path}")
+        else :
+            print("Le fichier existe dans data/son")
 
 ###la classe a deplacé
 
@@ -91,3 +95,6 @@ class gestion_dl():
         gestion_dl.dossier.mkdir(parents=True, exist_ok=True)
 
         print(f"Le répertoire '{gestion_dl.dossier}' a été créé avec succès.")
+
+    def verifier_dl(self, nom):
+
